+++ examples/.eslintrc.js
@@ -2,6 +2,7 @@ const path = require('path');
 
 module.exports = {
   extends: path.join(__dirname, '..', '.eslintrc.js'),
+  ignorePatterns: ['dist', 'build', '.next', '.nuxt'],
   rules: {
     'import/no-unresolved': 'off',
     'import/named': 'off',