+++ packages/instantsearch.js/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [4.90.0](https://github.com/algolia/instantsearch/compare/instantsearch.js@4.89.0...instantsearch.js@4.90.0) (2026-02-25)
+
+
+### Features
+
+* **build:** unify build system across monorepo ([#6886](https://github.com/algolia/instantsearch/issues/6886)) ([e8b2ef7](https://github.com/algolia/instantsearch/commit/e8b2ef7dc1b7d98f1ee5518fad7607b5f31bbd96))
+
+
+
+
+
+# [4.89.0](https://github.com/algolia/instantsearch/compare/instantsearch.js@4.88.0...instantsearch.js@4.89.0) (2026-02-24)
+
+
+### Features
+
+* **chat:** add compatibility with algolia mcp search tool [DASH-2294] ([#6904](https://github.com/algolia/instantsearch/issues/6904)) ([adf1037](https://github.com/algolia/instantsearch/commit/adf1037eb4f7088e2972212917ed2ac32b23072a))
+
+
+
+
+
 # [4.88.0](https://github.com/algolia/instantsearch/compare/instantsearch.js@4.87.2...instantsearch.js@4.88.0) (2026-02-17)